Uses batch for uploads/deletions
All the upload/deletion request use the same route on the server.
This route then figure out if it's a multiple files operation to call `batch`.
If it's not, the regular method is call.

Also, the routes that doesn't respond with a body use a `204` status to
prevent from `XML parsing error` on the browser.