Unifile modal window creator
This creates a modal window with an iframe inside to show the login page form
Te schema is: a main layer,  that cover the whole screen, contain a layer smaller and inside this one there are three zones: title, content and footer. 
Th tittle contains title of the windows and a small close button 
The content contain the iframe for login page an footer contain a close button.