ModalDialog for login
Added section and layers for a login window, that includes am iframe with the login web page. Includes two function
One, setwin, it inits a state value (winOb) with the url of the window in the iframe.
Other getwin, for reading the actual value of the Url of the winfows in the iframe and close if the page is callback page, because in this moment the login form has been sent.